Rename packages/files and fix nits
