+++ core-api/src/main/java/com/optimizely/ab/event/internal/payload/Event.java
@@ -14,6 +14,6 @@
  *    See the License for the specific language governing permissions and
  *    limitations under the License.
  */
-package com.optimizely.ab.event.internal.payloadV2;
+package com.optimizely.ab.event.internal.payload;
 
-public class V2Event { }
+public class Event { }
