+++ core-api/src/main/java/com/optimizely/ab/event/internal/payload/LayerState.java
@@ -14,7 +14,7 @@
  *    See the License for the specific language governing permissions and
  *    limitations under the License.
  */
-package com.optimizely.ab.event.internal.payloadV2;
+package com.optimizely.ab.event.internal.payload;
 
 public class LayerState {
