+++ core-api/src/main/java/com/optimizely/ab/event/internal/serializer/GsonSerializer.java
@@ -18,13 +18,13 @@ package com.optimizely.ab.event.internal.serializer;
 
 import com.google.gson.Gson;
 
-import com.optimizely.ab.event.internal.payloadV2.V2Event;
+import com.optimizely.ab.event.internal.payload.Event;
 
 class GsonSerializer implements Serializer {
 
     private Gson gson = new Gson();
 
-    public <T extends V2Event> String serialize(T payload) {
+    public <T extends Event> String serialize(T payload) {
         return gson.toJson(payload);
     }
 }
