+++ core-api/src/main/java/com/optimizely/ab/event/internal/serializer/Serializer.java
@@ -16,8 +16,8 @@
  */
 package com.optimizely.ab.event.internal.serializer;
 
-import com.optimizely.ab.event.internal.payloadV2.V2Event;
+import com.optimizely.ab.event.internal.payload.Event;
 
 public interface Serializer {
-    <T extends V2Event> String serialize(T payload) throws SerializationException;
+    <T extends Event> String serialize(T payload) throws SerializationException;
 }
