+++ core-api/src/test/java/com/optimizely/ab/event/internal/serializer/JacksonSerializerTest.java
@@ -16,8 +16,8 @@
  */
 package com.optimizely.ab.event.internal.serializer;
 
-import com.optimizely.ab.event.internal.payloadV2.Conversion;
-import com.optimizely.ab.event.internal.payloadV2.Impression;
+import com.optimizely.ab.event.internal.payload.Conversion;
+import com.optimizely.ab.event.internal.payload.Impression;
 
 import org.junit.Test;
