+++ core-api/src/test/resources/serializer/conversion.json
@@ -31,8 +31,6 @@
       "value": 5000
     }
   ],
-  "eventFeatures": [
-
-  ],
+  "eventFeatures": [],
   "isGlobalHoldback": false
 }
